Move stuff to config
