+++ server/controllers/dashboard/dashboard.controller.ts
@@ -1,14 +1,20 @@
 import { FastifyRequest, FastifyReply } from 'fastify';
-const DATABASE_URL: string = "http://database_container:3000";
+import { User, UserPayload } from '../../types';
+import { DATABASE_URL, LOCAL_GAMESERVER_URL, USERMANAGEMENT_URL } from '../../config';
+
 
 export async function getDashboard(request: FastifyRequest, reply: FastifyReply) {
+	// get UserPayload from cookie, this is managed by the JWT server decoration.
+	const payload: UserPayload = request.user;
+
+	console.log("payload", payload)
+
   try {
-    if (request.headers['hx-request']) {
-      return reply.viewAsync("dashboard/dashboard-view.ejs");
-    }
-    // return reply.sendFile("dashboard.html");
-    return reply.viewAsync("dashboard/dashboard-view.ejs");
-  }
+		return reply.viewAsync("dashboard/dashboard-view.ejs", {
+			user: payload.name,
+			gameserverURL: LOCAL_GAMESERVER_URL
+		});
+	}
   catch (error) {
     request.log.error(error);
     return reply.viewAsync("errors/error-500.ejs");
@@ -16,6 +22,7 @@ export async function getDashboard(request: FastifyRequest, reply: FastifyReply)
 }
 
 export async function getDashboardUser(request: FastifyRequest, reply: FastifyReply) {
+
   const { username } = request.params as { username: string };
 
   try {
